requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
gunicorn==21.2.0
firebase-admin==6.4.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
